# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Flowstile API
  version: 0.1.0
  description: >-
    Human-task inbox and form layer for Temporal.io workflows.


    Authentication uses JWT tokens delivered as HttpOnly cookies (`flowstile_token`) or as Bearer tokens in the
    Authorization header.
  license:
    name: Apache-2.0
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowstile_token
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /cases:
    get:
      tags:
        - Cases
      parameters:
        - schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
          in: query
          name: limit
          required: false
        - schema:
            default: 0
            type: integer
            minimum: 0
            maximum: 9007199254740991
          in: query
          name: offset
          required: false
        - schema:
            type: string
            enum:
              - pending
              - in_progress
              - completed
              - cancelled
          in: query
          name: status
          required: false
      responses:
        "200":
          description: Default Response
  /cases/by-process-instance/{processInstanceId}:
    get:
      tags:
        - Cases
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: processInstanceId
          required: true
      responses:
        "200":
          description: Default Response
  /cases/{id}:
    get:
      tags:
        - Cases
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /cases/by-process-instance/{processInstanceId}/variables:
    patch:
      tags:
        - Cases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                variables:
                  type: object
                  propertyNames:
                    type: string
                  additionalProperties: {}
              required:
                - variables
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: processInstanceId
          required: true
      responses:
        "200":
          description: Default Response
security:
  - cookieAuth: []
  - bearerAuth: []
...